feat: rework Select to use reactstrap implementation
docs have still to be updated but examples work
