+++ stories/Form/Select.stories.mdx
@@ -41,32 +41,6 @@ Per disabilitare una select, aggiungere la _prop_ `isDisabled` al componente `Se
   <Story id='componenti-form-select--select-disabilitata' />
 </Canvas>
 
-## Select con reset
-
-Per creare l’opzione che resetterà i valori selezionati andremo ad inserire all'interno dei componenti personalizzati della `Select` un componente `ClearIndicator`.
-
-In aggiunta imposteremo l'attributo `isClearable={true}` per indicare che il campo può tornare nello stato iniziale.
-
-<Canvas>
-  <Story id='componenti-form-select--select-con-reset' />
-</Canvas>
-
-## Select con ricerca
-
-Per attivare la ricerca è sufficiente aggiungere l'attributo `isSearchable={true}` al componente `Select`.
-
-<Canvas>
-  <Story id='componenti-form-select--select-con-ricerca' />
-</Canvas>
-
-## Select Multipla
-
-Per attivare la multiselezione, aggiungere l'attributo `isMulti={true}` al componente `Select`.
-
-<Canvas>
-  <Story id='componenti-form-select--select-multipla' />
-</Canvas>
-
 ## Select con gruppi
 
 Per le select con raggruppamenti il formato di input dei dati è leggermente diversa dal formato precedente:
@@ -102,22 +76,6 @@ In aggiunta al formato diverso è necessario passare due componenti `GroupHeadin
 
 <Canvas>
   <Story id='componenti-form-select--select-con-gruppi' />
-</Canvas>
-
-## Typescript
-
-Di seguito alcuni esempi di uso del componente `Select` con i tipi Typescript.
-
-### Select classica
-
-<Canvas>
-  <Story id='componenti-form-select--select-classica-ts' />
-</Canvas>
-
-### Select con Gruppi
-
-<Canvas>
-  <Story id='componenti-form-select--select-con-gruppi-ts' />
 </Canvas>
 
 ## Argomenti componente
